+++ packages/expo-ui/CHANGELOG.md
@@ -11,12 +11,15 @@
 - Adds `DateTimePicker` component. ([#34883](https://github.com/expo/expo/pull/34883) by [@alanjhughes](https://github.com/alanjhughes))
 - Add CircularProgress and LinearProgress component. ([#34907](https://github.com/expo/expo/pull/34907) by [@janicduplessis](https://github.com/janicduplessis))
 - Add Gauge component ([#35032](https://github.com/expo/expo/pull/35032) by [@jakex7](https://github.com/jakex7))
+- Add List and Label component ([#](https://github.com/expo/expo/pull/) by [@Pflaumenbaum](https://github.com/Pflaumenbaum))
 
 ### 🐛 Bug fixes
 
 ### 💡 Others
 
 - Standardize platform key ordering in `expo-module.config.json`. ([#35003](https://github.com/expo/expo/pull/35003) by [@reichhartd](https://github.com/reichhartd))
+- Updated UnwrappedChildren to return a ForEach type, enabling support for modifiers like onDelete. ([#](https://github.com/expo/expo/pull/) by [@Pflaumenbaum](https://github.com/Pflaumenbaum))
+
 
 ## 0.0.2 — 2025-02-11